refactor: 클래스 이름 변경 및 기능 축소
- 클래스 이름을 JwtProvider로 변경
- 이름에 맞게 토큰 생성과 정보 조회 기능만 남김
